+++ eth_defi/event_reader/multithread.py
@@ -1,12 +1,12 @@
 """Multithreaded and parallel Solidity event reading helpers."""
-from typing import Any, Optional, List, Iterable, Counter
+from typing import Any, Optional, List, Iterable, Counter, Callable
 
 from requests.adapters import HTTPAdapter
 from web3.contract.contract import ContractEvent
 
 from eth_defi.event_reader.filter import Filter
 from eth_defi.event_reader.logresult import LogResult
-from eth_defi.event_reader.reader import Web3EventReader, read_events_concurrent, ReaderConnection, ProgressUpdate
+from eth_defi.event_reader.reader import Web3EventReader, read_events_concurrent, ReaderConnection, ProgressUpdate, read_events
 from eth_defi.event_reader.reorganisation_monitor import ReorganisationMonitor
 from eth_defi.event_reader.web3factory import TunedWeb3Factory
 from eth_defi.event_reader.web3worker import create_thread_pool_executor
@@ -116,18 +116,49 @@ class MultithreadEventReader(Web3EventReader):
         reader.close()
         assert len(feeds) == 2
 
+    Because Ethereum does not have native JSON-RPC API to get block timestamps and headers
+    easily, there are many work arounds how to get timestamps for events.
+    Here is an example how to fetch timestamps "lazily" only for blocks
+    where you have events:
+
+    .. code-block:: python
+
+        from eth_defi.event_reader.lazy_timestamp_reader import extract_timestamps_json_rpc_lazy
+
+        provider = cast(HTTPProvider, web3.provider)
+        json_rpc_url = provider.endpoint_uri
+        reader = MultithreadEventReader(json_rpc_url, max_threads=16)
+
+        start_block = 1
+        end_block = web3.eth.block_number
+
+        reader = MultithreadEventReader(
+            json_rpc_url,
+        )
+
+        for log_result in reader(
+            web3,
+            restored_start_block,
+            end_block,
+            filter=filter,
+            extract_timestamps=extract_timestamps_json_rpc_lazy,
+        ):
+            pass
+
+    See :py:func:`eth_defi.event_reader.lazy_timestamp_reader.extract_timestamps_json_rpc_lazy`
+    and :py:func:`eth_defi.uniswap_v3.events.fetch_events_to_csv` for more details.
     """
 
     def __init__(
-        self,
-        json_rpc_url: str,
-        max_threads=10,
-        reader_context: Any = None,
-        api_counter=True,
-        max_blocks_once=50_000,
-        reorg_mon: Optional[ReorganisationMonitor] = None,
-        notify: Optional[ProgressUpdate] = None,
-        auto_close_notify=True,
+            self,
+            json_rpc_url: str,
+            max_threads=10,
+            reader_context: Any = None,
+            api_counter=True,
+            max_blocks_once=50_000,
+            reorg_mon: Optional[ReorganisationMonitor] = None,
+            notify: Optional[ProgressUpdate] = None,
+            auto_close_notify=True,
     ):
         """Creates a multithreaded JSON-RPC reader pool.
 
@@ -183,6 +214,9 @@ class MultithreadEventReader(Web3EventReader):
 
         # Set up the timestamp reading method
 
+    def get_max_threads(self) -> int:
+        return self.executor.max_workers
+
     def close(self):
         """Release the allocated resources."""
         self.executor.join()
@@ -193,12 +227,13 @@ class MultithreadEventReader(Web3EventReader):
                 self.notify.close()
 
     def __call__(
-        self,
-        web3: ReaderConnection,
-        start_block: int,
-        end_block: int,
-        events: Optional[List[ContractEvent]] = None,
-        filter: Optional[Filter] = None,
+            self,
+            web3: ReaderConnection,
+            start_block: int,
+            end_block: int,
+            events: Optional[List[ContractEvent]] = None,
+            filter: Optional[Filter] = None,
+            extract_timestamps: Optional[Callable] = None,
     ) -> Iterable[LogResult]:
         """Wrap the underlying low-level function.
 
@@ -208,21 +243,62 @@ class MultithreadEventReader(Web3EventReader):
 
             Currently timestamp reading not supported
 
+        :param web3:
+            Currently unused
+
+        :param start_block:
+            First block to call in eth_getLogs. Inclusive.
+
+        :param end_block:
+            End block to call in eth_getLogs. Inclusive.
+
+        :param events:
+            Event signatures we are interested in.
+
+            Legacy. Use ``filter`` instead.
+
+        :param filter:
+            Event filter we are using.
+
+        :param extract_timestamps:
+            Use this method to get timestamps for our events.
+
+            Overrides :py:attr:`reorg_mon` given in the constructor (if any given).
+            See usage examples in :py:class:`MultithreadEventReader`.
+
         :return:
             Iterator for the events in the order they were written in the chain
 
         """
-        yield from read_events_concurrent(
-            self.executor,
-            start_block,
-            end_block,
-            events=events,
-            filter=filter,
-            reorg_mon=self.reorg_mon,
-            notify=self.notify,
-            extract_timestamps=None,
-            chunk_size=self.max_blocks_once,
-        )
+
+        if self.get_max_threads() == 1:
+            # Single thread debug mode
+            web3 = self.web3_factory()
+            yield from read_events(
+                web3,
+                start_block,
+                end_block,
+                events=events,
+                filter=filter,
+                reorg_mon=self.reorg_mon,
+                notify=self.notify,
+                extract_timestamps=extract_timestamps,
+                chunk_size=self.max_blocks_once,
+            )
+
+        else:
+            # Multi thread production mode
+            yield from read_events_concurrent(
+                self.executor,
+                start_block,
+                end_block,
+                events=events,
+                filter=filter,
+                reorg_mon=self.reorg_mon,
+                notify=self.notify,
+                extract_timestamps=extract_timestamps,
+                chunk_size=self.max_blocks_once,
+            )
 
     def get_total_api_call_counts(self) -> Counter:
         """Sum API call counts across all threads.