+++ eth_defi/token.py
@@ -4,11 +4,13 @@ Deploy ERC-20 tokens to be used within your test suite.
 
 `Read also unit test suite for tokens to see how ERC-20 can be manipulated in pytest <https://github.com/tradingstrategy-ai/web3-ethereum-defi/blob/master/tests/test_token.py>`_.
 """
+from collections import OrderedDict
 from dataclasses import dataclass
 from decimal import Decimal
 from functools import cached_property
 from typing import Optional, Union
 
+import cachetools
 from eth_tester.exceptions import TransactionFailed
 from eth_typing import HexAddress
 from web3 import Web3
@@ -23,6 +25,12 @@ from eth_defi.utils import sanitise_string
 #: TODO: Add exceptios from real HTTPS/WSS providers
 #: `ValueError` is raised by Ganache
 _call_missing_exceptions = (TransactionFailed, BadFunctionCallOutput, ValueError, ContractLogicError)
+
+
+#: By default we cache 1024 token details using LRU.
+#:
+#:
+DEFAULT_TOKEN_CACHE = cachetools.LRUCache(1024)
 
 
 @dataclass
@@ -121,6 +129,19 @@ class TokenDetails:
         raw_amount = self.contract.functions.balanceOf(address).call(block_identifier=block_identifier)
         return self.convert_to_decimals(raw_amount)
 
+    @staticmethod
+    def generate_cache_key(chain_id: int, address: str) -> int:
+        """Generate a cache key for this token.
+
+        - Cached by (chain, address) tuple
+
+        - Validate the inputs before generating the key
+        """
+        assert type(chain_id) == int
+        assert type(address) == str
+        assert address.startswith("0x")
+        return hash((chain_id, address.lower()))
+
 
 class TokenDetailError(Exception):
     """Cannot extract token details for an ERC-20 token for some reason."""
@@ -168,6 +189,8 @@ def fetch_erc20_details(
     max_str_length: int = 256,
     raise_on_error=True,
     contract_name="ERC20MockDecimals.json",
+    cache: cachetools.Cache | None = DEFAULT_TOKEN_CACHE,
+    chain_id: int = None,
 ) -> TokenDetails:
     """Read token details from on-chain data.
 
@@ -199,16 +222,46 @@ def fetch_erc20_details(
     :param contract_name:
         Contract ABI file to use.
 
-        The default is `ERC20MockDecimals.json`. For USDC use `centre/FiatToken.json`.
+        The default is ``ERC20MockDecimals.json``. For USDC use ``centre/FiatToken.json``.
+
+    :param cache:
+        Use this cache for cache token detail calls.
+
+        The main purpose is to easily reduce JSON-RPC API call count.
+
+        By default, we use LRU cache of 1024 entries.
+
+        Set to ``None`` to disable the cache.
+
+        Instance of :py:class:`cachetools.Cache'.
+        See `cachetools documentation for details <https://cachetools.readthedocs.io/en/latest/#cachetools.LRUCache>`__.
+
+    :param chain_id:
+        Chain id hint for the cache.
+
+        If not given do ``eth_chainId`` RPC call to figure out.
 
     :return:
         Sanitised token info
     """
 
-    # No risk here, because we are not sending a transaction
-    token_address = Web3.to_checksum_address(token_address)
+    if not chain_id:
+        chain_id = web3.eth.chain_id
+
     erc_20 = get_deployed_contract(web3, contract_name, token_address)
 
+    key = TokenDetails.generate_cache_key(chain_id, token_address)
+
+    if cache is not None:
+        cached = cache.get(key)
+        if cached is not None:
+            return TokenDetails(
+                erc_20,
+                cached["name"],
+                cached["symbol"],
+                cached["supply"],
+                cached["decimals"],
+            )
     try:
         symbol = sanitise_string(erc_20.functions.symbol().call()[0:max_str_length])
     except _call_missing_exceptions as e:
@@ -249,4 +302,24 @@ def fetch_erc20_details(
             raise TokenDetailError(f"Token {token_address} missing totalSupply") from e
         supply = None
 
-    return TokenDetails(erc_20, name, symbol, supply, decimals)
+    token_details = TokenDetails(erc_20, name, symbol, supply, decimals)
+    if cache is not None:
+        cache[key] = {
+            "name": name,
+            "symbol": symbol,
+            "supply": supply,
+            "decimals": decimals,
+        }
+    return token_details
+
+
+def reset_default_token_cache():
+    """Purge the cached token data.
+
+    See :py:data:`DEFAULT_TOKEN_CACHE`
+    """
+    global DEFAULT_TOKEN_CACHE
+    # Cache has a horrible API
+    DEFAULT_TOKEN_CACHE.__dict__["_LRUCache__order"] = OrderedDict()
+    DEFAULT_TOKEN_CACHE.__dict__["_Cache__currsize"] = 0
+    DEFAULT_TOKEN_CACHE.__dict__["_Cache__data"] = dict()