+++ tests/enzyme/conftest.py
@@ -18,9 +18,13 @@ from web3.contract import Contract
 from eth_defi.provider.anvil import AnvilLaunch, launch_anvil
 from eth_defi.chain import install_chain_middleware
 from eth_defi.deploy import deploy_contract
+from eth_defi.provider.multi_provider import create_multi_provider_web3
 from eth_defi.token import create_token
 from eth_defi.trace import assert_transaction_success_with_explanation
 from eth_defi.uniswap_v2.deployment import deploy_uniswap_v2_like, UniswapV2Deployment, deploy_trading_pair
+
+
+logger = logging.getLogger(__name__)
 
 
 @pytest.fixture()
@@ -49,15 +53,13 @@ def anvil(request: FixtureRequest) -> AnvilLaunch:
             log_level = log_cli_level
 
     # London hardfork will enable EIP-1559 style gas fees
-    anvil = launch_anvil(
-        hardfork="london",
-        gas_limit=15_000_000,  # Max 5M gas per block, or per transaction in test automining
-    )
+    anvil = launch_anvil()
     try:
         # Make the initial snapshot ("zero state") to which we revert between tests
         # web3 = Web3(HTTPProvider(anvil.json_rpc_url))
         # snapshot_id = make_anvil_custom_rpc_request(web3, "evm_snapshot")
         # assert snapshot_id == "0x0"
+        logger.info("Anvil launched at %s", anvil.json_rpc_url)
         yield anvil
     finally:
         anvil.close(log_level=log_level)
@@ -69,13 +71,7 @@ def web3(anvil: AnvilLaunch) -> Web3:
 
     Also perform the Anvil state reset for each test.
     """
-    web3 = Web3(HTTPProvider(anvil.json_rpc_url, request_kwargs={"timeout": 2}))
-
-    # Get rid of attributeddict slow down
-    web3.middleware_onion.clear()
-
-    install_chain_middleware(web3)
-
+    web3 = create_multi_provider_web3(anvil.json_rpc_url)
     return web3