+++ tests/uniswap_v2/test_uniswap_v2_synthetic.py
@@ -8,7 +8,7 @@ from eth_account.signers.local import LocalAccount
 from hexbytes import HexBytes
 from web3 import EthereumTesterProvider, Web3
 
-from eth_defi.token import create_token, TokenDetails, fetch_erc20_details
+from eth_defi.token import create_token, TokenDetails, fetch_erc20_details, reset_default_token_cache
 from eth_defi.uniswap_v2.deployment import (
     UniswapV2Deployment,
     deploy_uniswap_v2_like,
@@ -31,6 +31,10 @@ def eth_tester(tester_provider):
 @pytest.fixture
 def web3(tester_provider):
     """Set up a local unit testing blockchain."""
+
+    # This test does not work with token cache
+    reset_default_token_cache()
+
     # https://web3py.readthedocs.io/en/stable/examples.html#contract-unit-tests-in-python
     return Web3(tester_provider)